refactor(providers): move theme context and useTheme into their own module
ThemeProvider.tsx exported the provider component next to the useTheme hook,
so Vite Fast Refresh dropped the resolved theme on every edit
(react-refresh/only-export-components).

ThemeContext, the Theme/ResolvedTheme/ThemeContextValue types and useTheme now
live in themeContext.ts; ThemeProvider.tsx exports only its component. The
providers barrel keeps exporting both, and the two direct importers (App,
ThemeToggle) now go through it.

Closes the last of the 6 Fast Refresh warnings — eslint . is clean.
