+++ src/providers/index.ts
@@ -1,1 +1,3 @@
-export { ThemeProvider, useTheme } from "./ThemeProvider";
+export { ThemeProvider } from "./ThemeProvider";
+export { useTheme } from "./themeContext";
+export type { ResolvedTheme, Theme, ThemeContextValue } from "./themeContext";
