+++ src/providers/themeContext.ts
@@ -1,0 +1,26 @@
+import { createContext, useContext } from "react";
+
+/**
+ * Theme context and its consumer hook. Kept apart from `ThemeProvider.tsx` so
+ * that module exports only its component and keeps its Fast Refresh state when
+ * edited.
+ */
+
+export type Theme = "light" | "dark" | "system";
+export type ResolvedTheme = "light" | "dark";
+
+export interface ThemeContextValue {
+  theme: Theme;
+  resolvedTheme: ResolvedTheme;
+  setTheme: (theme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextValue | undefined>(undefined);
+
+export function useTheme(): ThemeContextValue {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+}
