+++ src/shared/components/ui/ThemeToggle.tsx
@@ -1,6 +1,6 @@
 import { Monitor, Moon, Sun } from "lucide-react";
 import { useTranslation } from "react-i18next";
-import { useTheme } from "@/src/providers/ThemeProvider";
+import { useTheme } from "@/src/providers";
 
 export function ThemeToggle() {
   const { theme, resolvedTheme, setTheme } = useTheme();
